Idem...
git-svn-id: svn://localhost/qrezix/branches/qrezix--qt4/qrezix@827 30024bfc-89fa-0310-95ff-ee513f6fc940

+++ rzxserverlistener.cpp
@@ -175,7 +175,7 @@ void RzxServerListener::connectToXnetserver()
 	reconnection.stop();
 	
 	QString serverHostname = RzxConfig::serverName();
-	unsigned long serverPort = RzxConfig::serverPort();
+	quint16 serverPort = RzxConfig::serverPort();
 	if (serverHostname.isEmpty() || !serverPort) {
 		notify(tr("Server name and port are not configured"));
 		return;
@@ -263,7 +263,7 @@ void RzxServerListener::parse(const QString& msg) {
 		else return;
 			
 		bool ok;	
-		iconHost = RzxHostAddress::fromRezix(msgParams.stripWhiteSpace().toULong(&ok, 16));
+		iconHost = RzxHostAddress::fromRezix(msgParams.stripWhiteSpace().toUInt(&ok, 16));
 		if (!ok)
 			return;
 		iconMode = true;
